Move some of the Ambassador Cloud related code around
