+++ pkg/dnet/loopback.go
@@ -1,4 +1,4 @@
-package loopback
+package dnet
 
 import (
 	"errors"
